Fix a bug in calculating flesch score

+++ src/document/Document.java
@@ -166,7 +166,7 @@ public abstract class Document {
 	{
 	    // TODO: You will play with this method in week 1, and 
 		// then implement it in week 2
-	    return 206.835 - 1.015 * (getNumWords() / getNumSentences()) - 84.6 * (getNumSyllables() / getNumWords());
+	    return 206.835 - 1.015 * ((double)getNumWords() / getNumSentences()) - 84.6 * ((double)getNumSyllables() / getNumWords());
 	}
